healthmate_ai/tools/pdf_parser_tool.py
import os
from typing import Dict, Any
try:
    from pypdf import PdfReader
except ImportError:
    try:
        from PyPDF2 import PdfReader
    except ImportError:
        PdfReader = None

class PDFParserTool:
    def __init__(self):
        if PdfReader is None:
            raise ImportError("pypdf or PyPDF2 is required for PDFParserTool")

    def extract_text(self, file_path: str) -> str:
        if not os.path.exists(file_path):
            raise FileNotFoundError(f"File not found: {file_path}")
        
        try:
            reader = PdfReader(file_path)
            text = ""
            for page in reader.pages:
                text += page.extract_text() + "\n"
            return text
        except Exception as e:
            raise Exception(f"Failed to parse PDF: {str(e)}")

    def parse_report(self, file_path: str) -> Dict[str, Any]:
        """
        Extracts text and returns a structured dictionary.
        In a real scenario, this might use regex or LLM to structure the data.
        """
        raw_text = self.extract_text(file_path)
        return {
            "file_path": file_path,
            "raw_text": raw_text,
            # Placeholder for structured data extraction
            "extracted_fields": {} 
        }
