healthmate_ai/requirements.txt
google-generativeai
pypdf
pdfplumber
aiohttp
python-dotenv
jsonschema
